backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
boto3
pandas
